Dict. sladoc: function documentation with functions as keys.
Script "get_docstring.py" reads all "*.f" and "*.F" files and
extracts the comment in them. The part of the comment at the
beginning of each file, beginning with "*+" and ending with
"*-", is copied into a dictionary. This dictionary has function
names, in lowercase, as keys.

The "setup.py" script runs the "get_docstring.py" script and saves
the dictionary in "lib/docstring_pickle.pkl" pickle file. Then it
is copied into the package distribution.

The module name is now "pyslalib" with "slalib" as a sub-module.
The "__init__.py" file for the module is in "lib". When the
"pyslalib" module is loaded the dictionary is read in from the
"docstring_pickle.pkl" file and saved as "sladoc".

Example:

>>> from pyslalib import slalib, sladoc
>>> slalib.sla_veri()
205004
>>> print sladoc['sla_caldj']
.
.
.

File "docstring_pickle.pkl" added to .gitignore.